tests/wpt/web-platform-tests/css/work-in-progress/gadams/cssom/cssstylerule-interface.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - CSSStyleRule - Interface</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-cssstylerule-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true('CSSStyleRule' in window);

      }, '\'CSSStyleRule\' in window');

      test(function(){

        assert_true(CSSStyleRule instanceof Function);

      }, 'CSSStyleRule instanceof Function');

      test(function(){

        assert_true('prototype' in CSSStyleRule);

      }, '\'prototype\' in CSSStyleRule');

      test(function(){

        assert_equals(CSSStyleRule.prototype.toString(), '[object CSSStyleRulePrototype]');

      }, 'CSSStyleRule.prototype.toString() == \'[object CSSStyleRulePrototype]\'');

      test(function(){

        assert_true('constructor' in CSSStyleRule.prototype);

      }, '\'constructor\' in CSSStyleRule.prototype');

      test(function(){

        assert_equals(CSSStyleRule.prototype.constructor,CSSStyleRule);

      }, 'CSSStyleRule.prototype.constructor === CSSStyleRule');

    ]]>
    </script>
  </body>
</html>
